RAG_SchoolBook/rag_tamil_book/requirements.txt
streamlit
python-dotenv
pdfplumber
pillow
requests
openai
psycopg2-binary
pgvector
neo4j
pyvis
jinja2